Added confirmation when leaving Add Bean page
Related to Discussion #587. This change gives the user a chance to double check if they really want to leave the Add Bean page if they've made changes. This is done using a string comparison by stringifying the data after loading the page and when leaving - if there were no changes the strings should be the same, and the dialog box will consequently not be shown.

Added a new string which requires translation for the other locales.

Also fixed a small typo in the english locale.