Fix flaky BrewBrewingGraphComponent test.
During the tests, `this.settings` is not set. This was causing a failure in onDestroy and prevent the component from being cleaned up properly. This resulted in random failures in unrelated tests with the error:
```
Uncaught Error: Uncaught (in promise): TypeError: _this5.brewComponent.timer?.isTimerRunning is not a function TypeError: _this5.brewComponent.timer?.isTimerRunning is not a function at http://localhost:9876/_karma_webpack_/webpack:/src/components/brews/brew-brewing-graph/brew-brewing-graph.component.ts:1319:35
```