Dynamically change web view origin per platform
To retain all data in the web view IndexedDB when updating from Cordova
builds, it is necessary to retain the same origin for the app files the
Cordova version used.

Both the scheme and the hostname are different between iOS and Android,
but Capacitor only support changing the scheme natively per platform. As
a workaround, the configuration now dynamically parses its own command
line arguments to determine the platform that is currently being
configured. Until Capacitor exposes the current platform in a documented
way (see https://github.com/ionic-team/capacitor/issues/3976), this will
have to suffice.

This slightly changes the workflow. Now, it is not possible to run
"npx cap sync" anymore to sync all platforms. Instead, the platforms
must be synced separately, for example using "npx cap sync android" and
"npx cap sync ios". For convenience, the command "npm run capsync" is
also provided to sync all platforms.

Note that the project currently only contains an Android target. The
ios target must be added to the "npm run capsync" command once it is
part of the repository.