Fix getJSONFile() by using Capacitor file system API
The Cordova file system plugin is not able to read JSON files in my
testing on Android. The root cause seems to be documented in this issue:
https://github.com/ionic-team/capacitor/issues/1564

While the monky patch suggested in the issue might work, it's easier to
just migrate to the native Capacitor API instead. In the future, all
file system access calls should be migrated to prevent similar issues.