Migrate file read functions in uiFileHelper to Capacitor API
The Cordova file system plugin does not work properly in Capacitor while
also using Angular. Some promises just do not resolve. While there is a
monkey-patch for that out there, migrating away from the unsupported
plugin seems like the better option.

This was also a great opportunity to clean up the file read methods in
uiFileHelper a bit, centralizing a lot of the file handling.

File writes and other remaining uses of the Cordova plugin must be
removed in the next steps.