Remove all usages of cordova-plugin-file in production code
This commit pretty much re-writes the import functionality which allows
us to remove all usages of cordova-plugin-file in production code.

Now only references in test code remain.

Note that the plugin itself cannot be removed yet, as we have to remove
cordova-plugin-advanced-http and cordova-plugin-file-transfer first
because they have dependencies on cordova-file-plugin.