Enable full import/export from/to directory on Android using SAF
This adds two new options to the settings to import/export all data
(including attachment files) to a directory on Android. This now once
again allows users to import/export all their data without using adb.

This is achieved using a little bit of custom native code to access the
necessary Android APIs, which are commonly known as the
Storage Access Framework (SAF).

The current implementation is a little bit hacky still, as it mostly
moves the SAF directory contents to a temporary cache directory before
working on them for the import. A similar process is used for the
export.

While this is inefficient and temporarily takes up additional disk
space, this was the easiest way to keep the common import/export code
the same while still allowing this.