Override angular dependency of ag-virtual-scroll
ag-virtual-scroll requests Angular 17, but Angular 18 is installed. To
have made this work in the first place, the --legacy-peer-deps flag
was used. The issue with that is that the dependency conflict will
confuse any other developer on every 'npm install'.

This commit adds an override section that basically tells npm that
ag-virtual-scroll is compatible with the installed angular version.

This override can be removed again when the following issue is fixed:
https://github.com/ericferreira1992/ag-virtual-scroll/issues/27