Replace TSLint with ESLint
TSLint is outdated and Typescript linting features were added to ESLint
instead. Furthermore, the old configuration was not runnable anyway.

The current linter settings are already turning off a lot of the
recommended rules. However, most files are still full of warnings and
errors. It is to be expected that linter rules need to be tweaked
further to reduce the number of hard to fix errors.