Use Capacitor Splashscreen plugin
Now the Splashscreen will not hide after a fixed time, but rather only
when the app is ready. This prevents us from showing a blank, white
screen to the user before we are ready.

The navigation bar changes in the Android resource files are required
because otherwise the navigation bar icons will be drawn white on a
white background. For very old API levels we just bail out and draw a
black navigation bar.