Use default splashscreen background on Android API 32 and above
Starting with API level 32 the new OS splashscreen API is available and
automatically used. The splashscreen image we set for lower API levels
causes the splashscreen background to be pure white or pure black,
depending on the theme. With this style override, the API level 32
splashscreen falls back to the system default, which looks a little
better.