Use UUIDs to create unique file names
The previous file name generation algorithm used a counter to build
uniqe file names from the prefix. This inherently caused old file names
to be re-used when a file is deleted and a new file is created later.

This re-use had the potential to cause caching issues. It also caused
seemingly random new images being linked to old data entries which have
missing images, for example because the entries were imported from a
JSON file, but the images were not.

Changing the file name generation to use UUIDv4 solves this issue, as
the randomness pretty much guarantees that we will never re-use old file
names.

Closes #833