Optimize async-image and disable loading placeholder image by default
Before this commit, the async-image component always resolved every
image URL twice. Now it is only resolved once.

The loading image was also turned into an optional feature because most
of the time it was only visible for around 100ms. The visual flicker
introduced by showing an image for 100ms is much worse than just showing
a blank space for that time.

Furthermore, the error and async handling was also cleaned up a bit.