Fix android keyboard padding issue by using capawesome for edge-to-edge
The stock capacitor options for configuring edge-to-edge and keyboard
overlays on android seem to only allow two choices.

1) The keyboard overlaps the WebView, making components at the bottom
   of a page not interactable.
2) The keyboard doesn't overlap the WebView, but there is a huge gap of
   padding between the keyboard and the WebView.

Using @capawesome/capacitor-android-edge-to-edge-support instead and
disabling the corresponding capacitor options seems to fix this on the
devices I tested. It looks like now there is no extra padding and
components on the bottom of the page are still interactable.

Fixes #972
See #971