Clean up code in ThemeService
I started working on this assuming the root cause for #1003 lies in the
ThemeService not applying the theme correctly in all cases, but this is
not the case.

However, this refactored version is easier to read as the theme
selection is now clearly separated into deciding which theme to use and
applying that theme.