Fix Android status bar background not applying correctly on startup
This commit moves the creation of the hacky Android inset/status bar
overlay to plugin initialization, which ensures that it is available as
soon as possible. This also removes potential for race conditions and
edge cases where devices could be too fast or too slow to correctly
apply the correct color.

However, this "solution" is still as hacky as before. If at all
possible, the whole mechanism should be replaced with the more robust
mechanism provided by @capawesome/capacitor-android-edge-to-edge-support
starting with Capacitor 8. See #1006.

Closes #1003