Pin @babel/plugin-transform-class-static-block to version 7.27.1
This is done in preparation for the Capacitor 8 update, which always
includes a full 'npm update'.

Why pin this package to this specific version? Because running
'npm update' without this would break the build as it would generate
invalid JavaScript code. The following error is generated:

ReferenceError: must call super constructor before using 'this' in
                derived class constructor

The offending code snippet (at least one example of it) is:

constructor(uiStorage, uiHelper, uiLog) {
  var _superprop_getAdd = () => super.add,
    _this = this;
  super(uiStorage, uiHelper, uiLog, 'SETTINGS');
  // [...]

Obviously this is invalid, exactly as described in the error message.

There already is an issue in angular-cli:
https://github.com/angular/angular-cli/issues/31674

The related upstream pull request in Babel is:
https://github.com/babel/babel/pull/17443

Pinning the dependency is obviously not a long-term solution, but is
required to keep the app running while working on all the tech updates.