Overhaul edge-to-edge, status bar, and navigation bar handling
After the update to Capacitor 8 we can now use the SystemBars Capacitor
core plugin to handle edge-to-edge concerns on modern devices. Corner
cases are still handled by
@capawesome/capacitor-android-edge-to-edge-support for now, which allows
us to remove the custom inset handling in the native Android code.

The legacy status bar and navigation bar calls are still required, at
least on the Android side. Removing them causes errors on API level 29
and below.

@capgo/capacitor-navigation-bar is also removed as in my testing it is
not required anymore.

This commit was tested on Android API level 26, 28, 29, 30, and 36. It
worked correctly on all of them.