Perform 'npm update' (within existing semver ranges)
As the blockers for this have been resolved (either by pinning offending
versions or updating them separately before) we can now run a full
'npm update' within the defined semver ranges.

In particular, this updates Angular to the latest version in the v20
branch. With that, ag-virtual-scroll can be updated manually next.