Disable Capacitor SystemBars insetsHandling for Android
The builtin SystemBars insetsHandling causes excessive padding at the
top bar of the application on Android. As we use
@capawesome/capacitor-android-edge-to-edge-support to support
edge-to-edge on Android instead of the builtin plugin, we disable the
insetsHandling.

It tested this on both old and new Android emulator devices, but we'll
have to wait and see what happens when these changes hit real users on
real devices. There's still a possibility of stuff breaking in specific
Android versions and/or WebView versions.